+++ changelog.md
@@ -3,6 +3,7 @@
 ## Upcoming
 * Launch creepers now have a smaller explosion 
 * Launch creepers have now reduced particles
+* Fisher and Pearler zombies no longer use the item if in water
 
 ## 1.12.1
 * Spawning feature (renamed from Base) can now be disabled